Refactored UI code, added events
