+++ HucaresWF/src/HucaresWF/CameraTab.cs
@@ -1,0 +1,36 @@
+﻿using Hucares.Server.Client.Models;
+using HucaresWF.Properties;
+using System;
+using System.Threading.Tasks;
+using System.Windows.Forms;
+
+namespace HucaresWF
+{
+    public partial class HucaresForm : Form
+    {
+        private async Task UpdateCameraDataSource()
+        {
+            cameraTable.DataSource = await cameraClient.GetAllCameras();
+        }
+
+        private async void submitCam_Click(object sender, EventArgs e)
+        {
+            if (!Uri.IsWellFormedUriString(hostUrlField.Text, UriKind.Absolute))
+            {
+                var eventArgs = new ExceptionEventArgs { ExceptionMessage = Resources.Error_BadUri };
+                ExceptionEventHandler?.Invoke(this, eventArgs);
+                return;
+            };
+            var camObj = new CameraInfo()
+            {
+                HostUrl = hostUrlField.Text,
+                Latitude = (double)latField.Value,
+                Longitude = (double)longField.Value,
+                IsTrustedSource = isTrustedBox.Checked
+            };
+
+            await cameraClient.InsertCamera(camObj);
+            await UpdateCameraDataSource();
+        }
+    }
+}
