+++ HucaresWF/src/HucaresWF/ExceptionEventArgs.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace HucaresWF
+{
+    public class ExceptionEventArgs : EventArgs
+    {
+        public string ExceptionMessage { get; set; }
+    }
+}
