+++ HucaresWF/src/HucaresWF/ExceptionEventMessageBoxObserver.cs
@@ -1,0 +1,12 @@
+﻿using System.Windows.Forms;
+
+namespace HucaresWF
+{
+    public class ExceptionEventMessageBoxObserver : IExceptionEventObserver
+    {
+        public void HandleExceptionEvent(object sender, ExceptionEventArgs args)
+        {
+            MessageBox.Show(args.ExceptionMessage);
+        }
+    }
+}
