+++ HucaresWF/src/HucaresWF/HucaresForm.resx
