+++ HucaresWF/src/HucaresWF/IExceptionEventObservable.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace HucaresWF
+{
+    public interface IExceptionEventObservable
+    {
+        event EventHandler<ExceptionEventArgs> ExceptionEventHandler;
+
+        void InvokeHandler(ExceptionEventArgs eventArgs);
+    }
+}
