+++ HucaresWF/src/HucaresWF/IExceptionEventObserver.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace HucaresWF
+{
+    public interface IExceptionEventObserver
+    {
+        void HandleExceptionEvent(object sender, ExceptionEventArgs args);
+    }
+}
